src/requirements.txt
asgiref==3.3.4
Django==3.2.4
djangorestframework==3.12.4
gunicorn==20.1.0
mysql==0.0.3
mysqlclient==2.0.3
pytz==2021.1
sqlparse==0.4.1
